Colocate mutations with their usage for easier maintenance.
